Wed, Jul 3, 2024, 11:55 AM +05:30

+++ credit/credit.c
@@ -2,7 +2,7 @@
 #include <cs50.h>
 #include <stdio.h>
 
-int quotient(int dividend, int divisor);
+int quotient(long dividend, int divisor);
 int checksum(long number);
 
 int main(void)
@@ -31,9 +31,9 @@ int checksum(long number)
 }
 
 // Function to to return quotient with out decimal part
-int quotient(int dividend, int divisor)
+int quotient(long dividend, int divisor)
 {
     int remainder = (dividend % divisor);
-    int quotient = (dividend - remainder) / divisor;
+    long quotient = (dividend - remainder) / divisor;
     return quotient;
 }
